requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=1.5.0
sqlalchemy>=2.0.0
tqdm>=4.65.0
